# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Onion Spider API
  version: "1.0.0"
  description: |
    REST API for the Onion Spider dark-web crawler.

    Authentication is established by `POST /api/auth/login` in one of two
    explicit modes:

    - **`Authorization: Bearer <token>`** — for scripts and non-browser clients.
      Send `"mode": "bearer"`; the token is then returned in the response body
      and no session cookie is set.
    - **Session cookie (default)** — omit `mode` (or send `"mode": "cookie"`).
      The response sets `os_session`
      (HttpOnly, Secure, SameSite=Strict), which browsers send automatically.
      This is what the web dashboard uses, so no token is ever stored in
      JavaScript-reachable response data or storage.

    The header wins when both are present.

    Cookie-authenticated requests using an unsafe method (anything other than
    GET/HEAD/OPTIONS) must also echo the `os_csrf` cookie back in an
    `X-CSRF-Token` header, or the request is rejected with 403. Bearer-authenticated
    requests are exempt: a browser will not attach an `Authorization` header to a
    cross-origin request on its own, so there is nothing to forge.

    The token embeds a `token_version`; it is rejected once the version is bumped
    (password reset or `logout-all`). Authorization (admin vs user) is resolved
    from the database on every request, not from the token claims.

    Operational endpoints `/healthz`, `/readyz` and `/metrics` are served on the
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: os_session
      description: >
        Set by POST /api/auth/login in the default cookie mode. HttpOnly,
        Secure, SameSite=Strict.
        Unsafe methods additionally require the X-CSRF-Token header (see the
        API description).
  schemas:
...
    Message:
      type: object
      properties:
        message: { type: string }
...
paths:
...
  /api/auth/verify:
    get:
      deprecated: true
      summary: Legacy email-verification page for links issued before fragment-token migration
      description: New links use /verify-account#token=... so the token is never sent in an HTTP URL.
      parameters:
        - in: query
          name: token
          required: true
          schema: { type: string }
      responses:
        '200': { description: HTML confirmation page, content: { text/html: {} } }
        '400': { description: Invalid token }
    post:
      summary: Consume an email-verification token (JSON or form body)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { token: { type: string } }
      responses:
        '200':
          description: Success page, or JSON when Accept includes application/json
          content:
            text/html: {}
            application/json: { schema: { $ref: '#/components/schemas/Message' } }
        '400': { description: Token invalid, expired or already used }
...